smaller data types

+++ account/accountInfo.go
@@ -11,8 +11,8 @@ import (
 
 // InfoResponse ...
 type InfoResponse struct {
-	BitcoinBalance uint64         `json:"bitcoin_balance" validate:"min=0"`
-	CsgokeyBalance uint64         `json:"csgokey_balance" validate:"min=0"`
+	BitcoinBalance float64        `json:"bitcoin_balance" validate:"min=0"`
+	CsgokeyBalance uint32         `json:"csgokey_balance" validate:"min=0"`
 	TradeLinkURL   sql.NullString `json:"trade_link_url"`
 }
